ASVO-556: Delete: only delete job if command successful

+++ web/tao/models.py
@@ -472,7 +472,11 @@ class WorkflowCommand(models.Model):
     execution_comment = models.TextField(null=True, blank=True)
 
     def __unicode__(self):
-        return self.command
+        if self.job_id is not None:
+            jid = " on {0}".format(self.job_id.id)
+        else:
+            jid = ""
+        return u"{cmd}{jid}".format(cmd=self.command, jid=jid)
 
     def jobid(self):
         return self.job_id.pk
@@ -486,11 +490,13 @@ class WorkflowCommand(models.Model):
             prev = WorkflowCommand.objects.get(pk=self.id)
             if self.execution_status != prev.execution_status:
                 self.executed = datetime.now()
-                if self.command == self.JOB_OUTPUT_DELETE:
-                    job_id = self.job_id.pk
-                    self.job_id = None
-                    Job.objects.get(id=job_id).delete()
-                    self.execution_comment += 'Job %(job_id)s successfully deleted.' % locals()
+                if self.command == self.JOB_OUTPUT_DELETE and \
+                    self.execution_status == COMPLETED:
+                        job_id = self.job_id.pk
+                        self.job_id = None
+                        Job.objects.get(id=job_id).delete()
+                        self.execution_comment += u'{0}\nJob {job_id} successfully deleted.'.format(
+                            self.execution_comment, job_id=job_id).strip()
         super(WorkflowCommand, self).save(force_insert=force_insert, force_update=force_update, using=using,
                                           update_fields=update_fields)
